Step 11 - You need to determine whether the task being added already exists or not. If it doesn't exist, you will add it, and if it already exists, you will set it up for editing. You can use the findIndex() method to accomplish this.
findIndex is an array method that lets find the index of the first element in an array that satisfies a given testing function.

Declare a dataArrIndex variable using const and set it to the result of the findIndex() method applied to the taskData array. Utilize arrow syntax to provide a callback function with item as the parameter, and within the callback, check if the id property of item is equal to the id property of currentTask.

If the task exists, this returns the index, and if it doesn't exist, it returns -1.