Step 38 - You need to remove the task from the DOM using remove() and from the taskData array using splice().
splice() is an array method that modifies arrays by removing, replacing, or adding elements at a specified index, while also returning the removed elements. It can take up to three parameters: the first one is the mandatory index at which to start, the second is the number of items to remove, and the third is an optional replacement element.

Use the remove() method to remove the parentElement of the buttonEl from the DOM. Then use splice() to remove the task from the taskData array. Pass in dataArrIndex and 1 as the parameters of your splice().

dataArrIndex is the index to start and 1 is the number of items to remove.