Step 1 - In this project, you'll build a decimal and binary converter and learn about both number systems. You'll also learn about recursion by using it to perform the conversions.
All of the HTML and CSS for this project has been provided for you.

When you're ready to get started, use the .getElementById() method to get the input element with the id number-input, and store it in a variable called numberInput. Use the same method to get the button element with the id convert-btn and store it in a variable called convertBtn, and the h2 element with the id result and store it in a variable called result.