Step 4 - Now that your checkUserInput() function is set up for testing, you can use an event listener to call the function when users click the Convert button.
Chain the .addEventListener() method to convertBtn. The event listener should listen for click events and take a reference to the checkUserInput function as a callback. Remember that function references are not called with parentheses.

Once that's done, whenever you click the Convert button, the value of the number input should be logged to the console.